Updated dependencies to VSC 1.75
VSC 1.75 was released in January of 2023. Most people are more than
likely to be on a higher version than this.

If this turns out to be a problem, the previous version dependency can probably be kept. We should use https://code.visualstudio.com/api/references/activation-events#onLanguage in this case.